# openapi.yaml
openapi: 3.0.0
info:
  title: Wetteranalyse API
  description: API for the Wetteranalyse (Weather Analysis) dashboard.
  version: 1.0.0

paths:
  /api/live_weather:
    get:
      summary: Get Live Weather
      description: Fetches current weather data for a given geographical coordinate from an external API and adds a reverse-geocoded location name.
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude of the location.
          schema:
            type: number
            format: float
        - name: lon
          in: query
          required: true
          description: Longitude of the location.
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Successful response with live weather data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    example: false
                  latitude:
                    type: number
                    format: float
                  longitude:
                    type: number
                    format: float
                  station_name:
                    type: string
                    description: "Human-readable location name."
                    example: "Berlin, 10115"
                  temperature:
                    type: number
                    format: float
                  relative_humidity:
                    type: number
                    format: float
                  wind_speed_10m:
                    type: number
                    format: float
                  rain:
                    type: number
                    format: float
                  timestamp:
                    type: string
                    format: date-time

  /api/all_stations:
    get:
      summary: Get All Stations
      description: Retrieves a list of all available weather stations from the database.
      responses:
        '200':
          description: A list of all weather stations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  stations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Station'

  /api/nearest_stations:
    get:
      summary: Get Nearest Stations
      description: Finds the 5 nearest weather stations to a given geographical coordinate.
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude of the location.
          schema:
            type: number
            format: float
        - name: lon
          in: query
          required: true
          description: Longitude of the location.
          schema:
            type: number
            format: float
      responses:
        '200':
          description: A list of the 5 nearest stations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  stations:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Station'
                        - type: object
                          properties:
                            distance_km:
                              type: number
                              format: float
                              description: "Distance in kilometers from the provided coordinates."

  /api/historical_data:
    get:
      summary: Get Historical Data
      description: Retrieves aggregated historical weather data for a specific station over a date range.
      parameters:
        - name: station_id
          in: query
          required: true
          description: The ID of the weather station.
          schema:
            type: integer
        - name: start_date
          in: query
          required: true
          description: Start date for the data range (YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          description: End date for the data range (YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: aggregation
          in: query
          required: false
          description: The aggregation level.
          schema:
            type: string
            enum: [daily, monthly, yearly]
            default: yearly
      responses:
        '200':
          description: A dictionary containing the aggregation level and the resulting data rows.
          content:
            application/json:
              schema:
                type: object
                properties:
                  aggregation:
                    type: string
                  rows:
                    type: array
                    items:
                      type: object

  /api/chart_data:
    get:
      summary: Get Chart Data
      description: Retrieves data formatted for generating a chart, based on a metric, station, date range, and aggregation level.
      parameters:
        - name: station_id
          in: query
          required: true
          description: The ID of the weather station.
          schema:
            type: integer
        - name: metric
          in: query
          required: true
          description: The metric to be plotted.
          schema:
            type: string
            enum: [TMK, TXK, TNK, RSK, UPM]
        - name: aggregation
          in: query
          required: true
          description: The aggregation level.
          schema:
            type: string
            enum: [daily, monthly, yearly]
        - name: start_date
          in: query
          required: true
          description: Start date for the data range (YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          description: End date for the data range (YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Data formatted for use in a chart.
          content:
            application/json:
              schema:
                type: object
                properties:
                  metric:
                    type: string
                  metric_label:
                    type: string
                  station_id:
                    type: integer
                  aggregation:
                    type: string
                  start_date:
                    type: string
                    format: date
                  end_date:
                    type: string
                    format: date
                  rows:
                    type: array
                    items:
                      type: object
                      properties:
                        period:
                          type: string
                        value:
                          type: number
                          format: float

components:
  schemas:
    Station:
      type: object
      properties:
        STATIONS_ID:
          type: integer
        VON_DATUM:
          type: integer
          description: "Start date of records in YYYYMMDD format."
        BIS_DATUM:
          type: integer
          description: "End date of records in YYYYMMDD format."
        GEOBREITE:
          type: number
          format: float
          description: "Latitude"
        GEOLAENGE:
          type: number
          format: float
          description: "Longitude"
        STATIONSNAME:
          type: string